new_admin_panel_sprint_1/03_sqlite_to_postgres/requirements.txt
flake8==4.0
psycopg2-binary==2.9
pytest==7.1
python-dateutil==2.8
python-dotenv==0.20

